+++ lib/stash/wrapper/embargo.rb
@@ -24,6 +24,13 @@ module Stash
         self.start_date = start_date
         self.end_date = end_date
       end
+
+      # Creates a new `Embargo` instance of type {EmbargoType::NONE} with the current date as start and end.
+      # @return [Embargo]
+      def self.none
+        today = Date.today
+        new(type: EmbargoType::NONE, period: EmbargoType::NONE.value, start_date: today, end_date: today)
+      end
     end
 
   end