+++ CHANGES.md
@@ -1,3 +1,9 @@
+# 0.1.2 (31 March 2016)
+
+- Add convenience method `Embargo.none` to create a no-embargo embargo element
+- Make `StashAdministrative` (and thereby `StashWrapper`) default to `Embargo.none`
+  if no `Embargo` is provided
+
 # 0.1.1 (16 March 2016)
 
 - Fix gem metadata